backend/requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
dj-rest-auth==2.1.2
Django==3.1.4
django-allauth==0.42.0
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
Faker==5.3.0
idna==2.10
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.8
pycparser==2.20
PyJWT==2.0.0
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
